Introduced Tools ParentMenu in Rubberduck and moved NavigateMenuItem Orders into NavigateParentMenu.cs to match other ParentMenus
