+++ RetailCoder.VBE/UI/Command/MenuItems/ParentMenus/NavigateParentMenu.cs
@@ -11,4 +11,14 @@ namespace Rubberduck.UI.Command.MenuItems.ParentMenus
 
         public override int DisplayOrder { get { return (int)RubberduckMenuItemDisplayOrder.Navigate; } }
     }
+
+    public enum NavigationMenuItemDisplayOrder
+    {
+        CodeExplorer,
+        //ToDoExplorer,
+        RegexSearchReplace,
+        FindSymbol,
+        FindAllReferences,
+        FindImplementations
+    }
 }
