--- RetailCoder.VBE/UI/Command/MenuItems/ParentMenus/NavigationMenuItemDisplayOrder.cs
@@ -1,12 +1,0 @@
-﻿namespace Rubberduck.UI.Command.MenuItems.ParentMenus
-{
-    public enum NavigationMenuItemDisplayOrder
-    {
-        CodeExplorer,
-        ToDoExplorer,
-        RegexSearchReplace,
-        FindSymbol,
-        FindAllReferences,
-        FindImplementations
-    }
-}
