+++ RetailCoder.VBE/UI/Command/MenuItems/ParentMenus/RubberduckParentMenu.cs
@@ -15,10 +15,9 @@ namespace Rubberduck.UI.Command.MenuItems.ParentMenus
         UnitTesting,
         Refactorings,
         Navigate,
+        Tools,
         CodeInspections,
-        SourceControl,
         Settings,
         About,
-        RegexAssistant
     }
 }
