+++ RetailCoder.VBE/UI/Command/MenuItems/ToDoExplorerCommandMenuItem.cs
@@ -1,5 +1,4 @@
-﻿using System.Windows.Input;
-using Rubberduck.UI.Command.MenuItems.ParentMenus;
+﻿using Rubberduck.UI.Command.MenuItems.ParentMenus;
 
 namespace Rubberduck.UI.Command.MenuItems
 {
@@ -10,7 +9,7 @@ namespace Rubberduck.UI.Command.MenuItems
         {
         }
 
-        public override string Key { get { return "RubberduckMenu_TodoItems"; } }
-        public override int DisplayOrder { get { return (int)NavigationMenuItemDisplayOrder.ToDoExplorer; } }
+        public override string Key { get { return "ToolsMenu_TodoItems"; } }
+        public override int DisplayOrder { get { return (int)ToolsMenuItemDisplayOrder.ToDoExplorer; } }
     }
 }
